+++ CumulusMX/EcowittLogFile.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Text.RegularExpressions;
 
+using Org.BouncyCastle.Ocsp;
+
 namespace CumulusMX
 {
 	internal partial class EcowittLogFile
@@ -65,8 +67,7 @@ namespace CumulusMX
 					continue;
 				}
 
-
-				//cumulus.LogDebugMessage($"EcowittLogFile.DataParser: Preprocessing record {fields[0]}");
+				cumulus.LogDebugMessage($"EcowittLogFile.DataParser: Preprocessing record {fields[0]}");
 
 				var rec = new EcowittApi.HistoricData();
 
@@ -85,44 +86,58 @@ namespace CumulusMX
 				int varInt;
 				int idx;
 
-				if (FieldIndex.TryGetValue("Indoor Temperature", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.IndoorTemp = varDec;
-				if (FieldIndex.TryGetValue("Indoor Humidity", out idx) && int.TryParse(fields[idx], out varInt)) rec.IndoorHum = varInt;
-				if (FieldIndex.TryGetValue("Outdoor Temperature", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Temp = varDec;
-				if (FieldIndex.TryGetValue("Outdoor Humidity", out idx) && int.TryParse(fields[idx], out varInt)) rec.Humidity = varInt;
-				if (FieldIndex.TryGetValue("Dew Point", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DewPoint = varDec;
-				if (FieldIndex.TryGetValue("Feels Like", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.FeelsLike = varDec;
-				if (FieldIndex.TryGetValue("Wind", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WindSpd = varDec;
-				if (FieldIndex.TryGetValue("Gust", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WindGust = varDec;
-				if (FieldIndex.TryGetValue("Wind Direction", out idx) && int.TryParse(fields[idx], out varInt)) rec.WindDir = varInt;
-				if (FieldIndex.TryGetValue("ABS Pressure", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.StationPressure = varDec;
-				if (FieldIndex.TryGetValue("REL Pressure", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Pressure = varDec;
-				if (FieldIndex.TryGetValue("Solar Rad", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Solar = (int?) varDec;
-				if (FieldIndex.TryGetValue("UV-Index", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.UVI = varDec;
+				if (FieldIndex.TryGetValue("indoor temperature", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.IndoorTemp = varDec;
+				if (FieldIndex.TryGetValue("indoor humidity", out idx) && int.TryParse(fields[idx], out varInt)) rec.IndoorHum = varInt;
+				if (FieldIndex.TryGetValue("outdoor temperature", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Temp = varDec;
+				if (FieldIndex.TryGetValue("outdoor humidity", out idx) && int.TryParse(fields[idx], out varInt)) rec.Humidity = varInt;
+				if (FieldIndex.TryGetValue("dew point", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DewPoint = varDec;
+				if (FieldIndex.TryGetValue("feels like", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.FeelsLike = varDec;
+				if (FieldIndex.TryGetValue("wind", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WindSpd = varDec;
+				if (FieldIndex.TryGetValue("gust", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WindGust = varDec;
+				if (FieldIndex.TryGetValue("wind direction", out idx) && int.TryParse(fields[idx], out varInt)) rec.WindDir = varInt;
+				if (FieldIndex.TryGetValue("abs pressure", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.StationPressure = varDec;
+				if (FieldIndex.TryGetValue("rel pressure", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Pressure = varDec;
+				if (FieldIndex.TryGetValue("solar rad", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Solar = (int?) varDec;
+				if (FieldIndex.TryGetValue("uv-index", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.UVI = varDec;
 				// These fields 14,15,16 do not appear in the GW3000 log files :(
-				//if (FieldIndex.TryGetValue("Console Battery", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.ConsoleBattery = varDec;
-				//if (FieldIndex.TryGetValue("External Supply Battery", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.ExternalSupplyBattery = varDec;
-				//if (FieldIndex.TryGetValue("Charge", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Charge = varDec;
+				//if (FieldIndex.TryGetValue("console battery", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.ConsoleBattery = varDec;
+				//if (FieldIndex.TryGetValue("external supply battery", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.ExternalSupplyBattery = varDec;
+				//if (FieldIndex.TryGetValue("charge", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.Charge = varDec;
 
 
 				if (cumulus.Gw1000PrimaryRainSensor == 0)
 				{
 					// Tipping bucket
-					if (FieldIndex.TryGetValue("Hourly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainRate = varDec; // really this is hourly rain from the file
-					//if (FieldIndex.TryGetValue("Event Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.EventRain = varDec;
-					//if (FieldIndex.TryGetValue("Daily Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DailyRain = varDec;
-					//if (FieldIndex.TryGetValue("Weekly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WeeklyRain = varDec;
-					//if (FieldIndex.TryGetValue("Monthly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.MonthlyRain = varDec;
-					if (FieldIndex.TryGetValue("Yearly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainYear = varDec;
+					if (FieldIndex.TryGetValue("rain rate", out idx) && decimal.TryParse(fields[idx], invc, out varDec))
+					{
+						rec.RainRate = varDec;
+					}
+					else if (FieldIndex.TryGetValue("hourly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec))
+					{
+						rec.RainRate = varDec; // really this is hourly rain from the file
+					}
+					//if (FieldIndex.TryGetValue("event rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.EventRain = varDec;
+					//if (FieldIndex.TryGetValue("daily rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DailyRain = varDec;
+					//if (FieldIndex.TryGetValue("weekly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WeeklyRain = varDec;
+					//if (FieldIndex.TryGetValue("monthly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.MonthlyRain = varDec;
+					if (FieldIndex.TryGetValue("yearly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainYear = varDec;
 				}
 				else if (fields.Length >= 25)
 				{
 					// piezo rain
-					if (FieldIndex.TryGetValue("Piezo Hourly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainRate = varDec; // really this is hourly rain from the file
-					//if (FieldIndex.TryGetValue("Piezo Event Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.EventRain = varDec;
-					//if (FieldIndex.TryGetValue("Piezo Daily Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DailyRain = varDec;
-					//if (FieldIndex.TryGetValue("Piezo Weekly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WeeklyRain = varDec;
-					//if ((FieldIndex.TryGetValue("Piezo Monthly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.MonthlyRain = varDec;
-					if (FieldIndex.TryGetValue("Piezo Yearly Rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainYear = varDec;
+					if (FieldIndex.TryGetValue("piezo rate", out idx) && decimal.TryParse(fields[idx], invc, out varDec))
+					{
+						rec.RainRate = varDec;
+					}
+					else if (FieldIndex.TryGetValue("piezo hourly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec))
+					{
+						rec.RainRate = varDec; // really this is hourly rain from the file
+					}
+					//if (FieldIndex.TryGetValue("piezo event rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.EventRain = varDec;
+					//if (FieldIndex.TryGetValue("piezo daily rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.DailyRain = varDec;
+					//if (FieldIndex.TryGetValue("piezo weekly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.WeeklyRain = varDec;
+					//if ((FieldIndex.TryGetValue("piezo monthly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.MonthlyRain = varDec;
+					if (FieldIndex.TryGetValue("piezo yearly rain", out idx) && decimal.TryParse(fields[idx], invc, out varDec)) rec.RainYear = varDec;
 				}
 
 				//cumulus.LogDebugMessage($"EcowittLogFile.DataParser: Converting record {fields[0]} to MX units");
@@ -147,68 +162,68 @@ namespace CumulusMX
 						rec.DewPoint = MeteoLib.CToF(rec.DewPoint);
 						rec.FeelsLike = MeteoLib.CToF(rec.FeelsLike);
 					}
+				}
 
-					// convert wind to user units
-					if ((int) WindUnit != cumulus.Units.Wind)
+				// convert wind to user units
+				if ((int) WindUnit != cumulus.Units.Wind)
+				{
+					switch (WindUnit)
 					{
-						switch (WindUnit)
-						{
-							case WindUnits.ms:
-								rec.WindSpd = ConvertUnits.WindMSToUser(rec.WindSpd);
-								rec.WindGust = ConvertUnits.WindMSToUser(rec.WindGust);
-								break;
-							case WindUnits.mph:
-								rec.WindSpd = ConvertUnits.WindMPHToUser(rec.WindSpd);
-								rec.WindGust = ConvertUnits.WindMPHToUser(rec.WindGust);
-								break;
-							case WindUnits.kph:
-								rec.WindSpd = ConvertUnits.WindKPHToUser(rec.WindSpd);
-								rec.WindGust = ConvertUnits.WindKPHToUser(rec.WindGust);
-								break;
-							case WindUnits.knots:
-								rec.WindSpd = ConvertUnits.WindKnotsToUser(rec.WindSpd);
-								rec.WindGust = ConvertUnits.WindKnotsToUser(rec.WindGust);
-								break;
-						}
+						case WindUnits.ms:
+							rec.WindSpd = ConvertUnits.WindMSToUser(rec.WindSpd);
+							rec.WindGust = ConvertUnits.WindMSToUser(rec.WindGust);
+							break;
+						case WindUnits.mph:
+							rec.WindSpd = ConvertUnits.WindMPHToUser(rec.WindSpd);
+							rec.WindGust = ConvertUnits.WindMPHToUser(rec.WindGust);
+							break;
+						case WindUnits.kph:
+							rec.WindSpd = ConvertUnits.WindKPHToUser(rec.WindSpd);
+							rec.WindGust = ConvertUnits.WindKPHToUser(rec.WindGust);
+							break;
+						case WindUnits.knots:
+							rec.WindSpd = ConvertUnits.WindKnotsToUser(rec.WindSpd);
+							rec.WindGust = ConvertUnits.WindKnotsToUser(rec.WindGust);
+							break;
 					}
+				}
 
-					// convert rain to user units
-					if ((int) RainUnit != cumulus.Units.Rain)
+				// convert rain to user units
+				if ((int) RainUnit != cumulus.Units.Rain)
+				{
+					if (RainUnit == RainUnits.mm)
 					{
-						if (RainUnit == RainUnits.mm)
-						{
-							rec.RainRate = ConvertUnits.RainMMToUser(rec.RainRate);
-							//rec.EventRain = ConvertUnits.RainMMToUser(rec.EventRain);
-							//rec.DailyRain = ConvertUnits.RainMMToUser(rec.DailyRain);
-							//rec.WeeklyRain = ConvertUnits.RainMMToUser(rec.WeeklyRain);
-							//rec.MonthlyRain = ConvertUnits.RainMMToUser(rec.MonthlyRain);
-							rec.RainYear = ConvertUnits.RainMMToUser(rec.RainYear);
-						}
-						else
-						{
-							rec.RainRate = ConvertUnits.RainINToUser(rec.RainRate);
-							//rec.EventRain = ConvertUnits.RainINToUser(rec.EventRain);
-							//rec.DailyRain = ConvertUnits.RainINToUser(rec.DailyRain);
-							//rec.WeeklyRain = ConvertUnits.RainINToUser(rec.WeeklyRain);
-							//rec.MonthlyRain = ConvertUnits.RainINToUser(rec.MonthlyRain);
-							rec.RainYear = ConvertUnits.RainINToUser(rec.RainYear);
-						}
+						rec.RainRate = ConvertUnits.RainMMToUser(rec.RainRate);
+						//rec.EventRain = ConvertUnits.RainMMToUser(rec.EventRain);
+						//rec.DailyRain = ConvertUnits.RainMMToUser(rec.DailyRain);
+						//rec.WeeklyRain = ConvertUnits.RainMMToUser(rec.WeeklyRain);
+						//rec.MonthlyRain = ConvertUnits.RainMMToUser(rec.MonthlyRain);
+						rec.RainYear = ConvertUnits.RainMMToUser(rec.RainYear);
 					}
-
-					// convert pressure to user units
-					var cumPress = (cumulus.Units.Press == 0 || cumulus.Units.Press == 1) ? PressUnits.hPa : (PressUnits) cumulus.Units.Press;
-					if (PressUnit != cumPress)
+					else
 					{
-						if (PressUnit == PressUnits.hPa)
-						{
-							rec.StationPressure = ConvertUnits.PressMBToUser(rec.StationPressure);
-							rec.Pressure = ConvertUnits.PressMBToUser(rec.Pressure);
-						}
-						else
-						{
-							rec.StationPressure = ConvertUnits.PressINHGToUser(rec.StationPressure);
-							rec.Pressure = ConvertUnits.PressINHGToUser(rec.Pressure);
-						}
+						rec.RainRate = ConvertUnits.RainINToUser(rec.RainRate);
+						//rec.EventRain = ConvertUnits.RainINToUser(rec.EventRain);
+						//rec.DailyRain = ConvertUnits.RainINToUser(rec.DailyRain);
+						//rec.WeeklyRain = ConvertUnits.RainINToUser(rec.WeeklyRain);
+						//rec.MonthlyRain = ConvertUnits.RainINToUser(rec.MonthlyRain);
+						rec.RainYear = ConvertUnits.RainINToUser(rec.RainYear);
+					}
+				}
+
+				// convert pressure to user units
+				var cumPress = (cumulus.Units.Press == 0 || cumulus.Units.Press == 1) ? PressUnits.hPa : (PressUnits) cumulus.Units.Press;
+				if (PressUnit != cumPress)
+				{
+					if (PressUnit == PressUnits.hPa)
+					{
+						rec.StationPressure = ConvertUnits.PressMBToUser(rec.StationPressure);
+						rec.Pressure = ConvertUnits.PressMBToUser(rec.Pressure);
+					}
+					else
+					{
+						rec.StationPressure = ConvertUnits.PressINHGToUser(rec.StationPressure);
+						rec.Pressure = ConvertUnits.PressINHGToUser(rec.Pressure);
 					}
 				}
 
@@ -256,27 +271,78 @@ namespace CumulusMX
 			for (var i = 0; i < fields.Length; i++)
 			{
 				string cleanedHeader = unitRegex.Replace(fields[i], "").Trim();
-				FieldIndex[cleanedHeader] = i;
+				FieldIndex[cleanedHeader.ToLower()] = i; // Ecowitt tend to mess around with the case!
 			}
 
-			TempUnit = fields[FieldIndex["Indoor Temperature"]].EndsWith("C)") || fields[FieldIndex["Indoor Temperature"]].EndsWith("℃)") ? TempUnits.C : TempUnits.F;
+			if (FieldIndex.TryGetValue("indoor temperature", out int idx))
+			{
+				TempUnit = fields[idx].ToLower().EndsWith("c)") || fields[idx].EndsWith("℃)") ? TempUnits.C : TempUnits.F;
+			}
+			else
+			{
+				cumulus.LogErrorMessage("EcowittLogFile.HeaderParser: Unable to determine temperature units, defaulting to Cumulus units");
+				TempUnit = (TempUnits) cumulus.Units.Temp;
+			}
 
-			var wind = FieldIndex["Wind"];
+			if (FieldIndex.TryGetValue("wind", out idx))
+			{
 
-			if (fields[wind].EndsWith("m/s)")) WindUnit = WindUnits.ms;
-			else if (fields[wind].EndsWith("mph)")) WindUnit = WindUnits.mph;
-			else if (fields[wind].EndsWith("km/h)")) WindUnit = WindUnits.kph;
-			else if (fields[wind].EndsWith("knots)")) WindUnit = WindUnits.knots;
-			else WindUnit = 0;
+				if (fields[idx].ToLower().EndsWith("m/s)")) WindUnit = WindUnits.ms;
+				else if (fields[idx].ToLower().EndsWith("mph)")) WindUnit = WindUnits.mph;
+				else if (fields[idx].ToLower().EndsWith("km/h)")) WindUnit = WindUnits.kph;
+				else if (fields[idx].ToLower().EndsWith("knots)")) WindUnit = WindUnits.knots;
+				else WindUnit = (WindUnits) cumulus.Units.Wind;
+			}
+			else
+			{
+				cumulus.LogErrorMessage("EcowittLogFile.HeaderParser: Unable to determine wind units, defaulting to Cumulus units");
+				WindUnit = (WindUnits) cumulus.Units.Wind;
+			}
 
+			if (FieldIndex.TryGetValue("abs pressure", out idx))
+			{
+				if (fields[idx].ToLower().EndsWith("hpa)")) PressUnit = PressUnits.hPa;
+				else if (fields[idx].ToLower().EndsWith("inhg)")) PressUnit = PressUnits.inHg;
+				else if (fields[idx].ToLower().EndsWith("kpa)")) PressUnit = PressUnits.kPa;
+				else if (fields[idx].ToLower().EndsWith("mmhg)")) PressUnit = PressUnits.mmHg;
+				else PressUnit = cumulus.Units.Press switch
+				{
+					0 or 1 => PressUnits.hPa,
+					2 => PressUnits.inHg,
+					3 => PressUnits.kPa,
+					_ => PressUnits.hPa
+				};
+			}
+			else
+			{
+				cumulus.LogErrorMessage("EcowittLogFile.HeaderParser: Unable to determine pressure units, defaulting to Cumulus units");
+				PressUnit = cumulus.Units.Press switch
+				{
+					0 or 1 => PressUnits.hPa,
+					2 => PressUnits.inHg,
+					3 => PressUnits.kPa,
+					_ => PressUnits.hPa
+				};
+			}
 
-			var press = FieldIndex["ABS Pressure"];
-			if (fields[press].EndsWith("hPa)")) PressUnit = PressUnits.hPa;
-			else if (fields[press].EndsWith("inHg)")) PressUnit = PressUnits.inHg;
-			else if (fields[press].EndsWith("kPa)")) PressUnit = PressUnits.kPa;
-			else PressUnit = 0;
-
-			RainUnit = fields[FieldIndex["Hourly Rain"]].EndsWith("mm)") ? RainUnits.mm : RainUnits.inch;
+			if (FieldIndex.TryGetValue("hourly rain", out idx))
+			{
+				RainUnit = fields[idx].ToLower().EndsWith("mm)") ? RainUnits.mm : RainUnits.inch;
+			}
+			else if (FieldIndex.TryGetValue("piezo hourly rain", out idx))
+			{
+				RainUnit = fields[idx].ToLower().EndsWith("mm)") ? RainUnits.mm : RainUnits.inch;
+			}
+			else
+			{
+				cumulus.LogErrorMessage("EcowittLogFile.HeaderParser: Unable to determine rain units, defaulting to Cumulus units");
+				RainUnit = cumulus.Units.Rain switch
+				{
+					0 => RainUnits.mm,
+					1 => RainUnits.inch,
+					_ => RainUnits.mm
+				};
+			}
 		}